LayoutTests/imported/w3c/web-platform-tests/css/compositing/parsing/background-blend-mode-computed-multiple-expected.txt


FAIL Property background-blend-mode value 'normal' assert_equals: expected "normal" but got "normal normal normal"
FAIL Property background-blend-mode value 'multiply' assert_equals: expected "multiply" but got "multiply multiply multiply"
FAIL Property background-blend-mode value 'normal, luminosity' assert_equals: expected "normal, luminosity" but got "normal luminosity normal"
FAIL Property background-blend-mode value 'screen, overlay' assert_equals: expected "screen, overlay" but got "screen overlay screen"
FAIL Property background-blend-mode value 'color, saturation' assert_equals: expected "color, saturation" but got "color saturation color"
FAIL Property background-blend-mode value 'normal, luminosity, color' assert_equals: expected "normal, luminosity, color" but got "normal luminosity color"
FAIL Property background-blend-mode value 'screen, overlay, screen' assert_equals: expected "screen, overlay, screen" but got "screen overlay screen"
